k8s module ignores kube config file
The issue and suggested solution that worked here are explained at
https://github.com/kubernetes-client/python/issues/1333#issuecomment-992369110

It appears that the combination of using the deprecated k8s module and installing
Kubernetes client Python library as a dependency of openshift library causes the
problem. Since the openshift library by itself is not needed, we can pull in
Kubernetes library explicitly. Specifying the new Kubernetes module for Ansible
completes the fix.